Final polish before 2.0 release
* Set browser environment to allow browser globals
* Set jest and mocha environment for files matching `**/*.test.*`
* Disallow confusing globals. (`create-react-app` does the exact same thing.)
* Require imports to be first in the file, as they're hoisted and can
  create confusion if they're not first.

I've tested this in Monocle and on Web Register.

Monocle:
* `products` has 17 errors that can't be autofixed.
* `sales` has 10 errors that can't be autofixed.
* `shared-react` has 9 errors that can't be autofixed.

The errors look mostly pretty legit, but there are a few annoying import
ordering-related ones.

Web Register has 58 errors that can't be autofixed. Many are
console.logs in the same file, quite a few more are variables that could
be constant. Overall it doesn't look too hard to fix or exclude the
issues.